backend/requirements.txt
Django
psycopg2
django-restframework
django-cors-headers
djangorestframework
Pillow
